// packages/deciphony-renderer/src/numberNotation/enums/numberNotationSkinKeyEnum.d.ts
export declare enum NumberNotationSkinKeyEnum {
    Measure = "measure",
    Sharp = "sharp",
    Flat = "flat",
    Double_sharp = "double_sharp",
    Double_flat = "double_flat",
    Natural = "natural",
    Close_line = "close_line",
    linked_close_line = "linked_close_line",
    Single_barline = "single_barline",
    Double_barline = "double_barline",
    StartRepeat_barline = "startRepeat_barline",
    EndRepeat_barline = "endRepeat_barline",
    Dashed_barline = "dashed_barline",
    Final_barline = "final_barline",
    Start_end_repeat_barline = "start_end_repeat_barline",
    Dotted_barline = "dotted_barline",
    Reverse_barline = "reverse_barline",
    Heavy_barline = "heavy_barline",
    Heavy_double_barline = "heavy_double_barline",
    linked_single_barline = "linked_single_barline",
    linked_double_barline = "linked_double_barline",
    linked_startRepeat_barline = "linked_startRepeat_barline",
    linked_endRepeat_barline = "linked_endRepeat_barline",
    linked_dashed_barline = "linked_dashed_barline",
    linked_final_barline = "linked_final_barline",
    linked_start_end_repeat_barline = "linked_start_end_repeat_barline",
    linked_dotted_barline = "linked_dotted_barline",
    linked_reverse_barline = "linked_reverse_barline",
    linked_heavy_barline = "linked_heavy_barline",
    linked_heavy_double_barline = "linked_heavy_double_barline",
    C = "C",
    C_sharp = "C_sharp",
    D_flat = "D_flat",
    D = "D",
    E_flat = "E_flat",
    E = "E",
    F = "F",
    F_sharp = "F_sharp",
    G_flat = "G_flat",
    G = "G",
    A_flat = "A_flat",
    A = "A",
    B_flat = "B_flat",
    B = "B",
    C_flat = "C_flat",
    '1_1' = "1_1",
    '1_4' = "1_4",
    '2_4' = "2_4",
    '3_4' = "3_4",
    '4_4' = "4_4",
    '3_8' = "3_8",
    '6_8' = "6_8",
    AugmentationDot_1 = "augmentationDot_1",
    AugmentationDot_2 = "augmentationDot_2",
    AugmentationDot_3 = "augmentationDot_3",
    Number_0 = "number_0",
    Number_1 = "number_1",
    Number_2 = "number_2",
    Number_3 = "number_3",
    Number_4 = "number_4",
    Number_5 = "number_5",
    Number_6 = "number_6",
    Number_7 = "number_7",
    Number_X = "number_X",// 节奏音符（无音高）
    ReduceLine_1 = "reduceLine_1",
    ReduceLine_2 = "reduceLine_2",
    ReduceLine_3 = "reduceLine_3",
    ReduceLine_4 = "reduceLine_4",
    ReduceLine_5 = "reduceLine_5",
    ReduceLine_6 = "reduceLine_6",
    addLine = "addLine",// 简谱高/低音点下方的横线
    /** 八度点：音符上方=高八度，下方=低八度，每点表示一个八度 */
    OctaveDot = "octaveDot",
    Bracket = "bracket",
    Brace = "brace",
    Square = "square",
    Repeat_coda = "repeat_coda",
    Repeat_to_coda = "repeat_to_coda",
    Repeat_segno = "repeat_segno",
    Repeat_dc = "repeat_dc",
    Repeat_ds = "repeat_ds",
    Repeat_fine = "repeat_fine",
    Repeat_dc_al_fine = "repeat_dc_al_fine",
    Repeat_dc_al_coda = "repeat_dc_al_coda",
    Repeat_ds_al_fine = "repeat_ds_al_fine",
    Repeat_ds_al_coda = "repeat_ds_al_coda",
    Accent_above = "acccent_above",
    Staccato_above = "staccato_above",
    Staccatissimo_above = "staccato_above",
    Tenuto_above = "tenuto_above",
    Loure_above = "loure_above",
    Marcato_above = "marcato_above",
    Accent_staccato_above = "accent_staccato_above",
    Marcato_staccato_above = "marcato_staccato_above",
    Marcato_tenuto_above = "marcato_tenuto_above",
    Staccatissimo_stroke_above = "staccatissimo_stroke_above",
    Staccatissimo_wedge_above = "staccatissimo_wedge_above",
    Stress_above = "stress_above",
    Tenuto_accent_above = "tenuto_accent_above",
    Unstress_above = "unstress_above",
    Open = "open",
    Muted = "muted",
    Harmonic = "harmonic",
    Up_bow = "up_bow",
    Down_bow = "down_bow",
    Turn = "turn",
    InvertedTurn = "invertedTurn",
    Trill = "trill",
    ShortTrill = "shortTrill",
    Mordent = "mordent",
    Arpeggio = "arpeggio",
    Up_arpeggio = "up_arpeggio",
    Down_arpeggio = "down_arpeggio",
    Bracket_arpeggio = "bracket_arpeggio",
    Up_arpeggio_straight = "up_arpeggio_straight",
    Down_arpeggio_straight = "down_arpeggio_straight",
    Dyn_PPP = "dyn_ppp",
    Dyn_PP = "dyn_pp",
    Dyn_P = "dyn_p",
    Dyn_MP = "dyn_mp",
    Dyn_MF = "dyn_mf",
    Dyn_F = "dyn_f",
    Dyn_FF = "dyn_ff",
    Dyn_FFF = "dyn_fff",
    Dyn_FP = "dyn_fp",
    Dyn_PF = "dyn_pf",
    Dyn_SF = "dyn_sf",
    Dyn_SFZ = "dyn_sfz",
    Dyn_SFF = "dyn_sff",
    Dyn_SFFZ = "dyn_sffz",
    Dyn_SFP = "dyn_sfp",
    Dyn_RFZ = "dyn_rfz",
    Dyn_RF = "dyn_rf",
    Dyn_FZ = "dyn_fz"
}